requirements.txt
alembic==1.9.2
aniso8601==9.0.1
beautifulsoup4==4.11.1
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Migrate==4.0.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.0
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
greenlet==2.0.1
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
oauthlib==3.2.2
protobuf==4.21.12
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
